refactor: Move benchmarks into single file
