+++ README.md
@@ -22,6 +22,7 @@ Features:
 - Cross plattform (Linux, macOS, Windows).
 - It's FAST!
     Benchmarks show that a pretty big table with weird constraints is build in 4 _μs_, which is `4*1e^6`.
+    To run the benchmarks do a `cargo install cargo-criterion` and run `cargo criterion`.
 
 Comfy-table is written for the current `stable` Rust version.
 Older Rust versions may work, but aren't officially supported.
