+++ benches/build_tables.rs
@@ -1,8 +1,11 @@
 use criterion::{criterion_group, criterion_main, Criterion};
 
 use comfy_table::presets::UTF8_FULL;
+use comfy_table::Boundary::*;
+use comfy_table::ColumnConstraint::*;
 use comfy_table::*;
 
+/// Build the readme table
 fn build_readme_table() {
     let mut table = Table::new();
     table.load_preset(UTF8_FULL)
@@ -34,9 +37,46 @@ fn build_readme_table() {
     let _ = table.lines();
 }
 
-pub fn bench_build_readme_table(crit: &mut Criterion) {
+/// Create a dynamic 10x10 Table with width 400 and unevenly distributed content.
+/// On top of that, most of the columns have some kind of constraint.
+fn build_huge_table() {
+    let mut table = Table::new();
+    table
+        .load_preset(UTF8_FULL)
+        .set_content_arrangement(ContentArrangement::DynamicFullWidth)
+        .set_table_width(400)
+        .set_header(vec![1, 2, 3, 4, 5, 6, 7, 8, 9, 10]);
+
+    // Create a 10x10 grid
+    for row_index in 0..10 {
+        let mut row = Vec::new();
+        for column in 0..10 {
+            row.push("SomeWord ".repeat((column + row_index * 2) % 10))
+        }
+        table.add_row(row);
+    }
+
+    table.set_constraints(vec![
+        UpperBoundary(Fixed(20)),
+        LowerBoundary(Fixed(40)),
+        Absolute(Fixed(5)),
+        Absolute(Percentage(3)),
+        Absolute(Percentage(3)),
+        Boundaries {
+            lower: Fixed(30),
+            upper: Percentage(10),
+        },
+    ]);
+
+    // Build the table.
+    let _ = table.lines();
+}
+
+pub fn build_tables(crit: &mut Criterion) {
+    crit.bench_function("Huge table", |b| b.iter(|| build_huge_table()));
+
     crit.bench_function("Readme table", |b| b.iter(|| build_readme_table()));
 }
 
-criterion_group!(benches, bench_build_readme_table);
+criterion_group!(benches, build_tables);
 criterion_main!(benches);
